NO-ISSUE: test(playwright): redistribute coverage tests into domain files
Move tests from standalone coverage-gaps files into the existing test
files where they logically belong: user robots and self-delete into
api-v1-org-user, robot regenerate and team invitations into
api-v1-repo-features, quota sub-resources into api-v1-advanced-features,
and V2 blob/auth tests into api-v2. Delete the coverage-gaps files.
